edge/requirements.txt
greenlet==1.1.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.2.0
SQLAlchemy==1.4.39
zmq==0.0.0
